+++ code/enemies.py
@@ -1,11 +1,11 @@
 import abc
 import pygame
 import random
-import tileset
-import level as game
-from settings import *
-from entities import Entity
-from particles import Rock, Arrow, MagicMissile, PBoomerang, BombSmoke, Flame
+import code.tileset as tileset
+import code.level as game
+from code.settings import *
+from code.entities import Entity
+from code.particles import Rock, Arrow, MagicMissile, PBoomerang, BombSmoke, Flame
 
 
 # Known issue : Monster waits for STATE_ACTION to end before getting hurt
@@ -1180,7 +1180,8 @@ class Darknut(Enemy):
                 particle.affects_enemy = False
                 particle.collision_damage = 0
             elif (isinstance(particle, BombSmoke)
-                  and particle.hitbox.colliderect(self.hitbox)):
+                  and particle.hitbox.colliderect(self.hitbox)
+                  and not self.invulnerable):
                 self.state = STATE_STUN
                 self.stun_starting_time = pygame.time.get_ticks()
             elif (isinstance(particle, Flame)
