+++ code/settings.py
@@ -97,7 +97,7 @@ DOOR_EVENT_LABEL = "Event locked door"
 OPEN_DOORS_LABEL = 'open doors'
 # Graphics paths
 GRAPHICS_EXTENSION = '.png'
-GRAPHICS_PATH = '../graphics/'
+GRAPHICS_PATH = './graphics/'
 ENEMIES_PATH = GRAPHICS_PATH + 'enemies/'
 FONT_PATH = GRAPHICS_PATH + 'font/'
 HUD_PATH = GRAPHICS_PATH + 'hud/'
@@ -110,7 +110,7 @@ HUD_PERMA_PATH = HUD_PATH + 'hud_perma.png'
 BLACK_PATH = LEVELS_PATH + 'black.png'
 # Maps file paths
 MAPS_EXTENSION = '.csv'
-MAPS_PATH = '../map/'
+MAPS_PATH = './map/'
 MAPS_ENEMIES = '_Enemies'
 MAPS_ITEMS = '_Items'
 MAPS_LIMITS = '_Limits'
@@ -157,7 +157,7 @@ FAIRY_LABEL = 'Fairy'
 KEY_LABEL = 'Key'
 TRIFORCE_LABEL = 'Triforce'
 # Audio paths
-AUDIO_PATH = '../audio/'
+AUDIO_PATH = './audio/'
 THEME_OVERWORLD = AUDIO_PATH + 'Overworld.mp3'
 THEME_DUNGEON = AUDIO_PATH + 'Dungeon.mp3'
 SOUND_GAME_OVER = AUDIO_PATH + 'Game_Over.wav'
